requirements.txt
boto3==1.4.7
botocore==1.7.36
docutils==0.14
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
s3transfer==0.1.11
six==1.11.0
